feat: add vitest, remove jest
