+++ support-lambdas/bigquery-acquisitions-publisher/package.json
@@ -1,28 +1,28 @@
 {
-  "name": "bigquery-acquisitions-publisher",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "scripts": {
-    "test": "jest",
-    "build": "esbuild --bundle --platform=node --target=node20 --outfile=target/typescript/index.js typescript/index.ts"
-  },
-  "devDependencies": {
-    "@guardian/prettier": "^8.0.1",
-    "@types/aws-lambda": "^8.10.147",
-    "@types/jest": "^29.5.14",
-    "@types/node": "^22.13.4",
-    "esbuild": "^0.25.0",
-    "jest": "^29.7.0",
-    "prettier": "^3.5.1",
-    "tslib": "^2.6.2",
-    "typescript": "^5.7.3"
-  },
-  "dependencies": {
-    "@aws-sdk/client-ssm": "^3.750.0",
-    "@aws-sdk/credential-provider-node": "^3.750.0",
-    "@google-cloud/bigquery": "^7.9.2",
-    "aws-lambda": "^1.0.7",
-    "zod": "^3.24.2"
-  }
+	"name": "bigquery-acquisitions-publisher",
+	"version": "1.0.0",
+	"main": "index.js",
+	"license": "MIT",
+	"scripts": {
+		"test": "vitest --run",
+		"test-watch": "vitest",
+		"build": "esbuild --bundle --platform=node --target=node20 --outfile=target/typescript/index.js typescript/index.ts"
+	},
+	"devDependencies": {
+		"@guardian/prettier": "^8.0.1",
+		"@types/aws-lambda": "^8.10.147",
+		"@types/node": "^22.13.4",
+		"esbuild": "^0.25.0",
+		"prettier": "^3.5.1",
+		"tslib": "^2.6.2",
+		"typescript": "^5.7.3",
+		"vitest": "^3.0.7"
+	},
+	"dependencies": {
+		"@aws-sdk/client-ssm": "^3.750.0",
+		"@aws-sdk/credential-provider-node": "^3.750.0",
+		"@google-cloud/bigquery": "^7.9.2",
+		"aws-lambda": "^1.0.7",
+		"zod": "^3.24.2"
+	}
 }
